refactor: separated the file loading functions into a separate file
